backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==0.3.84
langchain-groq==0.2.1
langsmith==0.7.32
MarkupSafe==3.0.3
orjson==3.11.8
packaging==25.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.4
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
Werkzeug==3.1.8
xxhash==3.6.0
zstandard==0.25.0
